app: add /api/drink/view
This allows checking a drink by its barcode.

To allow adding a drink dynamically, drop Charitea Mate from the
database. Sure, we could simply use another drink, but this is also a
preparation for a later test that uses the umockdev recording which uses
that very same barcode. So simply drop the drink from the database and
create it dynamically via the new create_drink fixture.

Signed-off-by: Bastian Krause <basti@randomprojects.de>